To better understand the Parse opareators and functions are split
Parse becomes Deserialize and Serialize for the apropriate operator/function
